requirements.txt
fastapi
uvicorn
python-multipart
SQLAlchemy
PyMySQL
python-slugify
python-jose
scikit-learn
pandas